rename_change_csv/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage==1.30.0
gcsfs==0.6.2
pandas==1.2.0 
fsspec==0.8.5